refactor(suite-native): organize discreet text cmps
