+++ suite-native/atoms/src/DiscreetText/useDiscreetMode.ts
@@ -1,0 +1,17 @@
+import { useAtom } from 'jotai';
+
+import { atomWithUnecryptedStorage } from '@suite-native/storage';
+
+const isDiscreetModeOn = atomWithUnecryptedStorage<boolean>('isDiscreetModeOn', false);
+export const useDiscreetMode = () => {
+    const [isDiscreetMode, setIsDiscreetMode] = useAtom(isDiscreetModeOn);
+
+    const handleSetIsDiscreetMode = (value: boolean) => {
+        setIsDiscreetMode(value);
+    };
+
+    return {
+        isDiscreetMode,
+        setIsDiscreetMode: handleSetIsDiscreetMode,
+    };
+};
