+++ suite-native/atoms/src/index.ts
@@ -11,7 +11,9 @@ export * from './Input/SearchInput';
 export * from './Input/Input';
 export * from './CheckBox';
 export * from './Switch';
-export * from './DiscreetText';
+export * from './DiscreetText/DiscreetText';
+export * from './DiscreetText/DiscreetCanvas';
+export * from './DiscreetText/useDiscreetMode';
 export * from './ListItem/ListItem';
 export * from './Sheet/BottomSheet';
 export * from './Sheet/useBottomSheetAnimation';
